requirements.txt
hdbscan==0.8.26
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
matplotlib==3.1.2
numpy==1.18.1
pandas==0.25.3
python-dateutil==2.8.1
regex==2020.5.14
requests==2.23.0
requests-oauthlib==1.3.0
scikit-learn==0.23.2
scikit-image==0.17.2
scipy==1.4.1
seaborn==0.10.1
sklearn==0.0
tensorboard==1.15.0
torch==1.4.0
tokenizers==0.7.0
torchvision==0.5.0
transformers==2.10.0
tweepy==3.9.0
umap-learn==0.4.3
wandb==0.9.2